requirements.txt
# Core
streamlit==1.28.0
python-dotenv==1.0.0

# AI SDKs
openai==1.31.0
google-generativeai==0.3.2
anthropic==0.7.0

# Search API
google-search-results==2.4.2

# Office & PPT
python-pptx==0.6.21
Pillow==10.0.0

# Data & plotting
numpy==1.26.0
pandas==2.1.1
matplotlib==3.7.0
seaborn==0.12.0
plotly==5.17.0
yfinance==0.2.18

# Requests
requests==2.31.0
httpx==0.25.0

# Pydantic core for OpenAI compatibility
pydantic==2.12.2
pydantic-core==2.12.2
